Fix invalid insertion hint to underlying set container
The code in the add_segment function tries to get the prior of an
iterator that may be the first iterator of the underlying set container.
This is then used as an insertion hint which causes a segfault in some
standard container libraries.

The fix is to use cyclic_prior instead.

This is a fix for ticket #12872.